Add users to Chat state.
Now server.js send a message of type 'userJoined' to everyone when
someone connects successfully (see joinChat function). The
'userJoined' message sends the list of the users tracked by the
server. When chatty.js receives the message, it update the state of
the Chat component by calling setState. When a websocket connection
dies, the server sends to everyone a 'userLeft' message with the new
list of users. The client updates the state with the new user list in
the same way it does when a 'userJoined' message is received.